feat: refatorando para fazer upload de imagens no S3

+++ src/plan/plan.service.ts
@@ -11,6 +11,7 @@ import { UpdatePlanDto } from './dto/update-plan.dto';
 import { PlanEntity } from './entities/plan.entity';
 import { PlanImageEntity } from './entities/plan-image.entity';
 import { PlanProductService } from 'src/plan-product/plan-product.service';
+import { S3Service } from 'src/s3/s3.service';
 
 @Injectable()
 export class PlanService {
@@ -21,6 +22,7 @@ export class PlanService {
     private readonly planImageRepository: Repository<PlanImageEntity>,
     @Inject(forwardRef(() => PlanProductService))
     private readonly planProductService: PlanProductService,
+    private readonly s3Service: S3Service,
   ) {}
 
   async create(createPlanDto: CreatePlanDto) {
@@ -70,12 +72,19 @@ export class PlanService {
       where: { id: planId },
     });
 
-    await this.planImageRepository.delete({ plan });
+    if (!plan) {
+      throw new NotFoundException('Plano não encontrado com o ID informado');
+    }
 
-    const imageEntities = images.map((image) =>
-      this.planImageRepository.create({
-        image: image.buffer,
-        plan: plan,
+    await this.planImageRepository.delete({ plan });
+    const imageEntities = await Promise.all(
+      images.map(async (image) => {
+        const imageUrl = await this.s3Service.uploadImage(image);
+        const imageEntity = this.planImageRepository.create({
+          imageUrl: imageUrl,
+          plan: plan,
+        });
+        return imageEntity;
       }),
     );
 
@@ -87,12 +96,23 @@ export class PlanService {
       where: { id: planId },
     });
 
+    const imageEntities = await this.planImageRepository.findBy({ plan });
+    await Promise.all(
+      imageEntities.map((image) => this.s3Service.deleteImage(image.imageUrl)),
+    );
     await this.planImageRepository.delete({ plan });
   }
 
   async remove(id: string) {
+    const imageEntities = await this.planImageRepository.findBy({
+      plan: { id: id },
+    });
+
+    await Promise.all(
+      imageEntities.map((image) => this.s3Service.deleteImage(image.imageUrl)),
+    );
+
     await this.planImageRepository.delete({ plan: { id } });
-    await this.planProductService.remove(id);
     return await this.planRepository.delete(id);
   }
 }
