+++ src/product/product.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { In, Repository } from 'typeorm';
 import { CreateProductDto } from './dto/create-product.dto';
@@ -6,6 +6,7 @@ import { UpdateProductDto } from './dto/update-product.dto';
 import { ProductEntity } from './entities/product.entity';
 import { CategoryService } from 'src/category/category.service';
 import { ProductImageEntity } from './entities/product-image.entity';
+import { S3Service } from 'src/s3/s3.service';
 
 @Injectable()
 export class ProductService {
@@ -15,6 +16,7 @@ export class ProductService {
     @InjectRepository(ProductImageEntity)
     private readonly productImageRepository: Repository<ProductImageEntity>,
     private readonly categoryService: CategoryService,
+    private readonly s3Service: S3Service,
   ) {}
 
   async create(createProductDto: CreateProductDto) {
@@ -29,9 +31,21 @@ export class ProductService {
 
   async findAll() {
     return await this.productRepository.find({
-      relations: ['category'],
+      relations: ['category', 'images'],
       order: { name: 'ASC', createdAt: 'DESC' },
     });
+  }
+
+  async findAllForCustomer() {
+    const products = await this.productRepository.find({
+      relations: ['category', 'images'],
+      order: { name: 'ASC', createdAt: 'DESC' },
+    });
+
+    return products.map((product) => ({
+      ...product,
+      image: product.images[0] ? product.images[0] : null,
+    }));
   }
 
   async findOne(id: string) {
@@ -69,12 +83,19 @@ export class ProductService {
       where: { id: productId },
     });
 
-    await this.productImageRepository.delete({ product });
+    if (!product) {
+      throw new NotFoundException('Produto não encontrado com o ID informado');
+    }
 
-    const imageEntities = images.map((image) =>
-      this.productImageRepository.create({
-        image: image.buffer,
-        product: product,
+    await this.productImageRepository.delete({ product });
+    const imageEntities = await Promise.all(
+      images.map(async (image) => {
+        const imageUrl = await this.s3Service.uploadImage(image);
+        const imageEntity = this.productImageRepository.create({
+          imageUrl: imageUrl,
+          product: product,
+        });
+        return imageEntity;
       }),
     );
 
@@ -86,10 +107,22 @@ export class ProductService {
       where: { id: productId },
     });
 
+    const imageEntities = await this.productImageRepository.findBy({ product });
+    await Promise.all(
+      imageEntities.map((image) => this.s3Service.deleteImage(image.imageUrl)),
+    );
     await this.productImageRepository.delete({ product });
   }
 
   async remove(id: string) {
+    const imageEntities = await this.productImageRepository.findBy({
+      product: { id: id },
+    });
+
+    await Promise.all(
+      imageEntities.map((image) => this.s3Service.deleteImage(image.imageUrl)),
+    );
+
     await this.productImageRepository.delete({ product: { id } });
     return await this.productRepository.delete(id);
   }
